+++ CHANGES.rst
@@ -2,8 +2,30 @@ Changelog
 =========
 
 
-1.0a1 (unreleased)
-------------------
+1.2 (unreleased)
+----------------
+
+- Implement an adapter for extra holidays
+  [mpeeters]
+
+
+1.1 (2017-04-28)
+----------------
+
+- Release on internal egge server.
+
+- Update the compute_due_date method to handle working days
+  [mpeeters]
+
+- Add a class to manage working days for date calculation
+  [mpeeters]
+
+- Handle working days for additional delay
+  [mpeeters]
+
+
+1.0 (2017-04-28)
+----------------
 
 - Initial release.
   [sdelcourt]